Pruebas unitarias

+++ src/main/java/com/example/demo/entity/Movimiento.java
@@ -21,7 +21,7 @@ public class Movimiento {
     private TipoMovimiento tipo;
 
     private String categoria;
-    private BigDecimal monto;
+    private double monto;
 
     @ManyToOne
     @JoinColumn(name = "cuenta_id")
